third_party/blink/web_tests/html/sections/section-element-expected.txt

Various tests for the section element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<section> closes <p>:
PASS section1.parentNode.nodeName == "p" is false
<p> does not close <section>:
PASS p1.parentNode.nodeName is "SECTION"
<section> can be nested inside <section>:
PASS section3.parentNode.id is "section2"
Residual style:
PASS getWeight("section4") is "700"
PASS getWeight("span1") is "700"
FormatBlock:
PASS document.getElementById("span2").parentNode.nodeName is "SECTION"
PASS successfullyParsed is true

TEST COMPLETE
